+++ lib/components/common/bar_graph.dart
