+++ lib/components/common/stat_card.dart
