+++ lib/components/dialogs/confirm_discard_dialog.dart
