+++ lib/components/pet_profile/pet_profile_avatar.dart
