+++ tests/components/info_modal_tests.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:cheart/components/info_modal.dart';
+import 'package:cheart/components/common/info_modal.dart';
 import 'package:cheart/themes/cheart_theme.dart';
 
 // === Navigator Observer to detect dialog pops ===
