+++ tests/components/post_session_modal.tests.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:provider/provider.dart';
 
-import 'package:cheart/components/post_session_modal.dart';
+import 'package:cheart/components/session/post_session_modal.dart';
 import 'package:cheart/models/respiratory_session_model.dart';
 import 'package:cheart/providers/respiratory_rate_provider.dart';
 import 'package:cheart/utils/respiratory_constants.dart';
